pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.90.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-google-genai>=4.2.2",
    "langchain-huggingface>=1.2.2",
    "langchain-pinecone>=0.2.13",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pinecone-client>=6.0.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "torchvision>=0.26.0",
]
